Refactor database query execution by removing loader module and integrating dbutil.execute method across multiple files
