+++ libs/data/__init__.py
@@ -3,7 +3,6 @@
 
 - `libs.data.aggregate`: データ集計
 - `libs.data.initialization`: データベース初期化
-- `libs.data.loader`: SQLクエリ実行
 - `libs.data.lookup`: 情報取得
 - `libs.data.modify`: DBデータ操作
 - `libs.data.search`: DB検索
