+++ libs/data/search.py
@@ -7,7 +7,6 @@ from contextlib import closing
 from typing import TYPE_CHECKING
 
 import libs.global_value as g
-from libs.data import loader
 from libs.domain.score import GameResult
 from libs.utils import dbutil
 
@@ -27,7 +26,7 @@ def for_db_score(first_ts: float) -> list[GameResult]:
 
     """
     data: list[GameResult] = []
-    rows = loader.execute(
+    rows = dbutil.execute(
         "select * from result where ts >= :first_ts and source like :source",
         {"first_ts": str(first_ts), "source": f"{g.adapter.interface_type}_%"},
     )
